refactor(virtual-mcp): extract ConnectionDialogContent into its own file
add-connection-dialog.tsx mixed the modal shell (dialog chrome, create/clone/
connect handlers) with the ~500-line search/grid/infinite-scroll body. The
body only reads its own props and internal hooks, with no shared closures
over the shell's state, so it moves out unchanged.
